requirements.txt
astroid==1.6.5
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
Django==2.1.5
django-cors-headers==3.1.1
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-simplejwt==4.3.0
get==2019.4.13
html5lib==1.0.1
idna==2.8
lazy-object-proxy==1.3.1
lxml==4.4.1
Markdown==2.6.11
Pillow==5.2.0
post==2019.4.13
public==2019.4.13
PyJWT==1.6.4
pytz==2018.5
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
six==1.11.0
soupsieve==1.9.2
sqlparse==0.3.0
urllib3==1.25.3
webencodings==0.5.1
wrapt==1.10.11
